+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Improved METAR output
 - Improved runway positioning for SimBrief import
 - Improved handling of runway orientation
+- Flight plan import sets cruise altitude
 
 ## [2.4.0] 2026-09-05